+++ plugin/jetbrains/src/main/kotlin/com/sourceplusplus/sourcemarker/reporting/error-report.kt
@@ -1,0 +1,251 @@
+package com.sourceplusplus.sourcemarker.reporting
+
+import com.intellij.diagnostic.*
+import com.intellij.ide.DataManager
+import com.intellij.ide.plugins.PluginManagerCore
+import com.intellij.ide.plugins.PluginUtil
+import com.intellij.idea.IdeaLogger
+import com.intellij.notification.NotificationListener
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.application.ApplicationNamesInfo
+import com.intellij.openapi.application.ex.ApplicationInfoEx
+import com.intellij.openapi.diagnostic.*
+import com.intellij.openapi.diagnostic.SubmittedReportInfo.SubmissionStatus
+import com.intellij.openapi.extensions.PluginId
+import com.intellij.openapi.progress.*
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.SystemInfo
+import com.intellij.util.Consumer
+import com.intellij.util.io.decodeBase64
+import com.sourceplusplus.sourcemarker.PluginBundle
+import org.eclipse.egit.github.core.*
+import org.eclipse.egit.github.core.client.GitHubClient
+import org.eclipse.egit.github.core.service.IssueService
+import java.awt.Component
+import java.io.*
+import java.util.*
+
+/**
+ * Allows SourceMarker bugs to be anonymously reported to GitHub.
+ * Adapted from: https://github.com/JuliaEditorSupport/julia-intellij
+ *
+ * @since 0.1.2
+ * @author [Brandon Fergerson](mailto:bfergerson@apache.org)
+ */
+private object AnonymousFeedback {
+
+    private const val gitRepoUser = "sourceplusplus"
+    private const val gitRepo = "sourcemarker"
+
+    /**
+     * Makes a connection to GitHub. Checks if there is an issue that is a duplicate and based on this, creates either a
+     * new issue or comments on the duplicate (if the user provided additional information).
+     *
+     * @param environmentDetails Information collected by [getKeyValuePairs]
+     * @return The report info that is then used in [GitHubErrorReporter] to show the user a balloon with the link
+     * of the created issue.
+     */
+    fun sendFeedback(environmentDetails: MutableMap<String, String>): SubmittedReportInfo {
+        try {
+            val client = GitHubClient()
+                .setOAuth2Token(String("OTcwM2I4Mjg5N2IyMmExNWM5YWNlNTU2Mjc1M2Y2MTlmMGZhMmY4Yg==".decodeBase64()))
+            val repoID = RepositoryId(gitRepoUser, gitRepo)
+            val issueService = IssueService(client)
+            var newGibHubIssue = createNewGibHubIssue(environmentDetails)
+            val duplicate = findFirstDuplicate(newGibHubIssue.title, issueService, repoID)
+            var isNewIssue = true
+            if (duplicate != null) {
+                issueService.createComment(repoID, duplicate.number, generateGitHubIssueBody(environmentDetails, false))
+                newGibHubIssue = duplicate
+                isNewIssue = false
+            } else newGibHubIssue = issueService.createIssue(repoID, newGibHubIssue)
+            return SubmittedReportInfo(
+                newGibHubIssue.htmlUrl, PluginBundle.message(
+                    if (isNewIssue) "git.issue.text" else "git.issue.duplicate.text",
+                    newGibHubIssue.htmlUrl,
+                    newGibHubIssue.number.toLong()
+                ),
+                if (isNewIssue) SubmissionStatus.NEW_ISSUE else SubmissionStatus.DUPLICATE
+            )
+        } catch (e: Exception) {
+            return SubmittedReportInfo(
+                null,
+                PluginBundle.message("report.error.connection.failure"),
+                SubmissionStatus.FAILED
+            )
+        }
+    }
+
+    private fun findFirstDuplicate(uniqueTitle: String, service: IssueService, repo: RepositoryId): Issue? {
+        val searchParameters = HashMap<String, String>(2)
+        searchParameters[IssueService.FILTER_STATE] = IssueService.STATE_OPEN
+        return service.pageIssues(repo, searchParameters).flatten().firstOrNull { it.title == uniqueTitle }
+    }
+
+    private fun createNewGibHubIssue(details: MutableMap<String, String>) = Issue().apply {
+        val errorMessage = details.remove("error.message")?.takeIf(String::isNotBlank) ?: "Unspecified error"
+        title = PluginBundle.message("git.issue.title", details.remove("error.hash").orEmpty(), errorMessage)
+        details["title"] = title
+        body = generateGitHubIssueBody(details, true)
+        labels = listOf(Label().apply { name = "generated-bug" }) //todo: figure out why labels aren't working
+    }
+
+    private fun generateGitHubIssueBody(details: MutableMap<String, String>, includeStacktrace: Boolean) =
+        buildString {
+            val errorDescription = details.remove("error.description").orEmpty()
+            val stackTrace = details.remove("error.stacktrace")?.takeIf(String::isNotBlank) ?: "invalid stacktrace"
+            if (errorDescription.isNotEmpty()) append(errorDescription).appendln("\n\n----------------------\n")
+            for ((key, value) in details) append("- ").append(key).append(": ").appendln(value)
+            if (includeStacktrace)
+                appendln("<details><summary>Full StackTrace</summary>")
+                    .appendln("<pre><code>")
+                    .appendln(stackTrace)
+                    .appendln("</code></pre>\n</details>")
+        }
+}
+
+/**
+ * todo: description.
+ */
+class GitHubErrorReporter : ErrorReportSubmitter() {
+    override fun getReportActionText() = PluginBundle.message("report.error.to.plugin.vendor")
+    override fun submit(
+        events: Array<IdeaLoggingEvent>,
+        info: String?,
+        parent: Component,
+        consumer: Consumer<SubmittedReportInfo>
+    ): Boolean {
+        // TODO improve
+        val event = events.firstOrNull { it.throwable != null } ?: return false
+        return doSubmit(event, parent, consumer, info)
+    }
+
+    private fun doSubmit(
+        event: IdeaLoggingEvent,
+        parent: Component,
+        callback: Consumer<SubmittedReportInfo>,
+        description: String?
+    ): Boolean {
+        val dataContext = DataManager.getInstance().getDataContext(parent)
+        val bean = GitHubErrorBean(
+            event.throwable,
+            IdeaLogger.ourLastActionId,
+            description ?: "<No description>",
+            event.message ?: event.throwable.message.toString()
+        )
+        PluginUtil.getInstance().findPluginId(event.throwable)?.let { pluginId ->
+            PluginManagerCore.getPlugin(pluginId)?.let { ideaPluginDescriptor ->
+                if (!ideaPluginDescriptor.isBundled) {
+                    bean.pluginName = ideaPluginDescriptor.name
+                    bean.pluginVersion = ideaPluginDescriptor.version
+                }
+            }
+        }
+
+        (event.data as? AbstractMessage)?.let { bean.attachments = it.includedAttachments }
+        val project = CommonDataKeys.PROJECT.getData(dataContext)
+        val reportValues = getKeyValuePairs(
+            bean,
+            ApplicationInfoEx.getInstanceEx(),
+            ApplicationNamesInfo.getInstance()
+        )
+        val notifyingCallback = CallbackWithNotification(callback, project)
+        val task = AnonymousFeedbackTask(
+            project, PluginBundle.message(
+                "report.error.progress.dialog.text"
+            ), true, reportValues, notifyingCallback
+        )
+        if (project == null) task.run(EmptyProgressIndicator()) else ProgressManager.getInstance().run(task)
+        return true
+    }
+
+    internal class CallbackWithNotification(
+        private val consumer: Consumer<SubmittedReportInfo>,
+        private val project: Project?
+    ) : Consumer<SubmittedReportInfo> {
+        override fun consume(reportInfo: SubmittedReportInfo) {
+            consumer.consume(reportInfo)
+            if (reportInfo.status == SubmissionStatus.FAILED) ReportMessages.GROUP.createNotification(
+                ReportMessages.getErrorReport(),
+                reportInfo.linkText, NotificationType.ERROR, null
+            ).setImportant(false).notify(project)
+            else ReportMessages.GROUP.createNotification(
+                ReportMessages.getErrorReport(), reportInfo.linkText,
+                NotificationType.INFORMATION, NotificationListener.URL_OPENING_LISTENER
+            ).setImportant(false).notify(project)
+        }
+    }
+}
+
+/**
+ * Extends the standard class to provide the hash of the thrown exception stack trace.
+ */
+class GitHubErrorBean(
+    throwable: Throwable,
+    val lastAction: String?,
+    val description: String,
+    val message: String
+) {
+    val exceptionHash: String
+    val stackTrace: String
+
+    init {
+        val trace = throwable.stackTrace
+        exceptionHash = Arrays.hashCode(trace).toString()
+        val sw = StringWriter()
+        throwable.printStackTrace(PrintWriter(sw))
+        stackTrace = sw.toString()
+    }
+
+    var pluginName = ""
+    var pluginVersion = ""
+    var attachments: List<Attachment> = emptyList()
+}
+
+/**
+ * todo: description.
+ */
+private class AnonymousFeedbackTask(
+    project: Project?, title: String, canBeCancelled: Boolean,
+    private val params: MutableMap<String, String>,
+    private val callback: Consumer<SubmittedReportInfo>
+) : Task.Backgroundable(project, title, canBeCancelled) {
+    override fun run(indicator: ProgressIndicator) {
+        indicator.isIndeterminate = true
+        callback.consume(AnonymousFeedback.sendFeedback(params))
+    }
+}
+
+/**
+ * Collects information about the running IDEA and the error
+ */
+private fun getKeyValuePairs(
+    error: GitHubErrorBean,
+    appInfo: ApplicationInfoEx,
+    namesInfo: ApplicationNamesInfo
+): MutableMap<String, String> {
+    PluginManagerCore.getPlugin(PluginId.findId("com.sourceplusplus.sourcemarker"))?.run {
+        if (error.pluginName.isBlank()) error.pluginName = name
+        if (error.pluginVersion.isBlank()) error.pluginVersion = version
+    }
+    val params = mutableMapOf(
+        "error.description" to error.description,
+        "Plugin Name" to error.pluginName,
+        "Plugin Version" to error.pluginVersion,
+        "OS Name" to SystemInfo.OS_NAME,
+        "Java Version" to SystemInfo.JAVA_VERSION,
+        "App Name" to namesInfo.productName,
+        "App Full Name" to namesInfo.fullProductName,
+        "App Version name" to appInfo.versionName,
+        "Is EAP" to java.lang.Boolean.toString(appInfo.isEAP),
+        "App Build" to appInfo.build.asString(),
+        "App Version" to appInfo.fullVersion,
+        "Last Action" to (error.lastAction ?: "Unknown"),
+        "error.message" to error.message,
+        "error.stacktrace" to error.stackTrace,
+        "error.hash" to error.exceptionHash
+    )
+    for (attachment in error.attachments) params["Attachment ${attachment.name}"] = attachment.encodedBytes
+    return params
+}